refactoring filters and configurations into their own packages
