+++ src/main/java/no/ndla/taxonomy/filters/LogFilter.java
@@ -1,4 +1,4 @@
-package no.ndla.taxonomy;
+package no.ndla.taxonomy.filters;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
