+++ src/main/java/no/ndla/taxonomy/filters/ResponseCacheHeadersFilter.java
@@ -1,4 +1,4 @@
-package no.ndla.taxonomy.rest.v1;
+package no.ndla.taxonomy.filters;
 
 import org.springframework.core.annotation.Order;
 import org.springframework.stereotype.Component;
